requirements.txt
PyJWT
Flask
flask_sqlalchemy
alembic